polus-precompute-mesh-plugin/src/requirements.txt
imageio==2.8.0
pandas==1.0.3
numpy==1.18.2
bfio==2.0.1
javabridge==1.0.19
python_bioformats==1.5.2
scikit_image==0.17.2
open3d==0.9
trimesh==3.8.19
wheel==0.36.2
